Fixes #45 : links from notification emails are now properly handled
Those links were kinda special, because they are in fact redirected (301) to another location, and so we needed to parse the redirected URL, not the original one (some vital details like the target page number are missing from the original link).

This led to a change in the UrlParser API, since we sometimes need to perform a network call to follow the redirection.