Sorry for this commit, consider this branch as a WIP
I also had to bump targetSdkVersion to the last version because of a
bug in Android Studio : some methods on AppCompatActivity where not found
in the IDE.

It forced me to find a solution to the Bundle limit problem on API 25.
It was there all along : use the "loadAndCache" mechanism from the
existing SubscriptionHandler class. Only problem left : bound cache
size (otherwise it will lead to an OutOfMemoryError).

Work to do :

- Check on API 25+ emulator
- Fix AlertDialogs (they are displayed full screen...)
- Bound SubscriptionHandler cache size (probably using Guava Cache)
- ...